refactor: bell_type 추가 및 서비스 v2로 교체
