+++ src/main/java/com/debatetimer/domain/customize/BellType.java
@@ -1,0 +1,9 @@
+package com.debatetimer.domain.customize;
+
+public enum BellType {
+
+    AFTER_START,
+    BEFORE_END,
+    AFTER_END,
+    ;
+}
