+++ src/main/java/com/debatetimer/dto/customize/request/BellRequest.java
@@ -1,13 +1,17 @@
 package com.debatetimer.dto.customize.request;
 
 import com.debatetimer.domain.customize.Bell;
+import com.debatetimer.domain.customize.BellType;
+import jakarta.validation.constraints.NotNull;
 
 public record BellRequest(
+        @NotNull
+        BellType type,
         int time,
         int count
 ) {
 
     public Bell toDomain() {
-        return new Bell(time, count);
+        return new Bell(type, time, count);
     }
 }
