+++ src/main/java/com/debatetimer/service/customize/CustomizeService.java
@@ -1,15 +1,19 @@
 package com.debatetimer.service.customize;
 
 import com.debatetimer.domain.customize.CustomizeTable;
+import com.debatetimer.domain.customize.CustomizeTimeBox;
 import com.debatetimer.domain.member.Member;
 import com.debatetimer.dto.customize.request.CustomizeTableCreateRequest;
 import com.debatetimer.dto.customize.response.CustomizeTableResponse;
+import com.debatetimer.entity.customize.BellEntity;
 import com.debatetimer.entity.customize.CustomizeTableEntity;
 import com.debatetimer.entity.customize.CustomizeTimeBoxEntities;
 import com.debatetimer.entity.customize.CustomizeTimeBoxEntity;
+import com.debatetimer.repository.customize.BellRepository;
 import com.debatetimer.repository.customize.CustomizeTableRepository;
 import com.debatetimer.repository.customize.CustomizeTimeBoxRepository;
 import java.util.List;
+import java.util.stream.IntStream;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,21 +24,26 @@ public class CustomizeService {
 
     private final CustomizeTableRepository tableRepository;
     private final CustomizeTimeBoxRepository timeBoxRepository;
+    private final BellRepository bellRepository;
 
     @Transactional
     public CustomizeTableResponse save(CustomizeTableCreateRequest tableCreateRequest, Member member) {
         CustomizeTable table = tableCreateRequest.toTable(member);
-        CustomizeTableEntity savedTable = tableRepository.save(new CustomizeTableEntity(table));
+        List<CustomizeTimeBox> timeBoxes = tableCreateRequest.toTimeBoxList();
 
-        CustomizeTimeBoxEntities savedCustomizeTimeBoxes = saveTimeBoxes(tableCreateRequest, savedTable.toDomain());
-        return new CustomizeTableResponse(savedTable.toDomain(), savedCustomizeTimeBoxes);
+        CustomizeTableEntity savedTableEntity = tableRepository.save(new CustomizeTableEntity(table));
+        saveTimeBoxes(savedTableEntity, timeBoxes);
+        return new CustomizeTableResponse(savedTableEntity.toDomain(), timeBoxes);
     }
 
     @Transactional(readOnly = true)
     public CustomizeTableResponse findTable(long tableId, Member member) {
         CustomizeTableEntity tableEntity = tableRepository.getByIdAndMember(tableId, member);
-        CustomizeTimeBoxEntities timeBoxes = timeBoxRepository.findTableTimeBoxes(tableEntity);
-        return new CustomizeTableResponse(tableEntity.toDomain(), timeBoxes);
+        List<CustomizeTimeBoxEntity> timeBoxEntityList = timeBoxRepository.findAllByCustomizeTable(tableEntity);
+        List<BellEntity> bellEntityList = bellRepository.findAllByCustomizeTimeBoxIn(timeBoxEntityList);
+        CustomizeTimeBoxEntities timeBoxEntities = new CustomizeTimeBoxEntities(timeBoxEntityList, bellEntityList);
+
+        return new CustomizeTableResponse(tableEntity.toDomain(), timeBoxEntities.toDomain());
     }
 
     @Transactional
@@ -43,38 +52,47 @@ public class CustomizeService {
             long tableId,
             Member member
     ) {
-        CustomizeTableEntity existingTable = tableRepository.getByIdAndMember(tableId, member);
-        CustomizeTable renewedTable = tableCreateRequest.toTable(member);
-        existingTable.updateTable(renewedTable);
+        CustomizeTableEntity tableEntity = tableRepository.getByIdAndMember(tableId, member);
+        tableEntity.updateTable(tableCreateRequest.toTable(member));
 
-        timeBoxRepository.deleteAllByTable(existingTable.getId());
-        CustomizeTimeBoxEntities savedCustomizeTimeBoxes = saveTimeBoxes(tableCreateRequest, existingTable.toDomain());
-        return new CustomizeTableResponse(existingTable.toDomain(), savedCustomizeTimeBoxes);
+        bellRepository.deleteAllByTable(tableEntity.getId());
+        timeBoxRepository.deleteAllByTable(tableEntity.getId());
+        List<CustomizeTimeBox> timeBoxes = tableCreateRequest.toTimeBoxList();
+        saveTimeBoxes(tableEntity, timeBoxes);
+        return new CustomizeTableResponse(tableEntity.toDomain(), timeBoxes);
     }
 
     @Transactional
     public CustomizeTableResponse updateUsedAt(long tableId, Member member) {
         CustomizeTableEntity tableEntity = tableRepository.getByIdAndMember(tableId, member);
-        CustomizeTimeBoxEntities timeBoxes = timeBoxRepository.findTableTimeBoxes(tableEntity);
-        tableEntity.updateUsedAt();
+        List<CustomizeTimeBoxEntity> timeBoxEntityList = timeBoxRepository.findAllByCustomizeTable(tableEntity);
+        List<BellEntity> bellEntityList = bellRepository.findAllByCustomizeTimeBoxIn(timeBoxEntityList);
+        CustomizeTimeBoxEntities timeBoxEntities = new CustomizeTimeBoxEntities(timeBoxEntityList, bellEntityList);
 
-        return new CustomizeTableResponse(tableEntity.toDomain(), timeBoxes);
+        tableEntity.updateUsedAt();
+        CustomizeTable table = tableEntity.toDomain();
+        List<CustomizeTimeBox> timeBoxes = timeBoxEntities.toDomain();
+        return new CustomizeTableResponse(table, timeBoxes);
     }
 
     @Transactional
     public void deleteTable(long tableId, Member member) {
         CustomizeTableEntity table = tableRepository.getByIdAndMember(tableId, member);
+
+        bellRepository.deleteAllByTable(table.getId());
         timeBoxRepository.deleteAllByTable(table.getId());
         tableRepository.delete(table);
     }
 
-    private CustomizeTimeBoxEntities saveTimeBoxes(
-            CustomizeTableCreateRequest tableCreateRequest,
-            CustomizeTable table
-    ) {
-        CustomizeTimeBoxEntities customizeTimeBoxes = tableCreateRequest.toTimeBoxes(table);
-        List<CustomizeTimeBoxEntity> savedTimeBoxes = timeBoxRepository.saveAll(
-                customizeTimeBoxes.getTimeBoxes());
-        return new CustomizeTimeBoxEntities(savedTimeBoxes);
+    private void saveTimeBoxes(CustomizeTableEntity tableEntity, List<CustomizeTimeBox> timeBoxes) {
+        IntStream.range(0, timeBoxes.size())
+                .forEach(i -> saveTimeBox(tableEntity, timeBoxes.get(i), i + 1));
+    }
+
+    private void saveTimeBox(CustomizeTableEntity tableEntity, CustomizeTimeBox timeBox, int sequence) {
+        CustomizeTimeBoxEntity timeBoxEntity = timeBoxRepository.save(
+                new CustomizeTimeBoxEntity(tableEntity, timeBox, sequence));
+        timeBox.getBells()
+                .forEach(bell -> bellRepository.save(new BellEntity(timeBoxEntity, bell)));
     }
 }
