+++ src/main/resources/db/migration/V12__add_bell_type_into_bell.sql
@@ -1,0 +1,2 @@
+ALTER TABLE bell
+    ADD COLUMN bell_type enum ('AFTER_START','BEFORE_END','AFTER_END') NOT NULL
