+++ src/test/java/com/debatetimer/controller/customize/CustomizeDocumentTest.java
@@ -18,6 +18,7 @@ import com.debatetimer.controller.BaseDocumentTest;
 import com.debatetimer.controller.RestDocumentationRequest;
 import com.debatetimer.controller.RestDocumentationResponse;
 import com.debatetimer.controller.Tag;
+import com.debatetimer.domain.customize.BellType;
 import com.debatetimer.domain.customize.CustomizeBoxType;
 import com.debatetimer.domain.customize.Stance;
 import com.debatetimer.dto.customize.request.BellRequest;
@@ -71,8 +72,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                         fieldWithPath("table[].boxType").type(STRING).description("타임 박스 유형"),
                         fieldWithPath("table[].time").type(NUMBER).description("발언 시간(초)").optional(),
                         fieldWithPath("table[].bell").type(ARRAY).description("종소리 정보").optional(),
-                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)").optional(),
-                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수").optional(),
+                        fieldWithPath("table[].bell[].type").type(STRING).description("종소리 종류"),
+                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)"),
+                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수"),
                         fieldWithPath("table[].timePerTeam").type(NUMBER).description("팀당 발언 시간 (초)").optional(),
                         fieldWithPath("table[].timePerSpeaking").type(NUMBER).description("1회 발언 시간 (초)").optional(),
                         fieldWithPath("table[].speaker").type(STRING).description("발언자 이름").optional()
@@ -95,8 +97,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                         fieldWithPath("table[].boxType").type(STRING).description("타임 박스 유형"),
                         fieldWithPath("table[].time").type(NUMBER).description("발언 시간(초)").optional(),
                         fieldWithPath("table[].bell").type(ARRAY).description("종소리 정보").optional(),
-                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)").optional(),
-                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수").optional(),
+                        fieldWithPath("table[].bell[].type").type(STRING).description("종소리 종류"),
+                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)"),
+                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수"),
                         fieldWithPath("table[].timePerTeam").type(NUMBER).description("팀당 발언 시간 (초)").optional(),
                         fieldWithPath("table[].timePerSpeaking").type(NUMBER).description("1회 발언 시간 (초)").optional(),
                         fieldWithPath("table[].speaker").type(STRING).description("발언자 이름").optional()
@@ -109,9 +112,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                             "반대", true, true),
                     List.of(
                             new CustomizeTimeBoxCreateRequest(Stance.PROS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellRequest(90, 1)), null, null, "콜리"),
+                                    120, List.of(new BellRequest(BellType.AFTER_START, 90, 1)), null, null, "콜리"),
                             new CustomizeTimeBoxCreateRequest(Stance.CONS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellRequest(90, 1), new BellRequest(120, 2)), null, null, "비토"),
+                                    120, List.of(new BellRequest(BellType.AFTER_START, 90, 1), new BellRequest(BellType.AFTER_START, 120, 2)), null, null, "비토"),
                             new CustomizeTimeBoxCreateRequest(Stance.NEUTRAL, "난상 토론", CustomizeBoxType.TIME_BASED,
                                     null, null, 360, 120, null),
                             new CustomizeTimeBoxCreateRequest(Stance.NEUTRAL, "존중 토론", CustomizeBoxType.TIME_BASED,
@@ -124,9 +127,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                             "찬성", "반대", true, true),
                     List.of(
                             new CustomizeTimeBoxResponse(Stance.PROS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellResponse(90, 1)), null, null, "콜리"),
+                                    120, List.of(new BellResponse(BellType.AFTER_START, 90, 1)), null, null, "콜리"),
                             new CustomizeTimeBoxResponse(Stance.CONS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellResponse(90, 1), new BellResponse(120, 2)), null, null, "비토"),
+                                    120, List.of(new BellResponse(BellType.AFTER_START, 90, 1), new BellResponse(BellType.AFTER_START, 120, 2)), null, null, "비토"),
                             new CustomizeTimeBoxResponse(Stance.NEUTRAL, "난상 토론", CustomizeBoxType.TIME_BASED,
                                     null, null, 360, 120, null),
                             new CustomizeTimeBoxResponse(Stance.NEUTRAL, "존중 토론", CustomizeBoxType.TIME_BASED,
@@ -168,9 +171,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                             "반대", true, true),
                     List.of(
                             new CustomizeTimeBoxCreateRequest(Stance.PROS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellRequest(90, 1)), null, null, "콜리"),
+                                    120, List.of(new BellRequest(BellType.AFTER_START, 90, 1)), null, null, "콜리"),
                             new CustomizeTimeBoxCreateRequest(Stance.CONS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellRequest(90, 1), new BellRequest(120, 2)), null, null, "비토"),
+                                    120, List.of(new BellRequest(BellType.AFTER_START, 90, 1), new BellRequest(BellType.AFTER_START, 120, 2)), null, null, "비토"),
                             new CustomizeTimeBoxCreateRequest(Stance.NEUTRAL, "난상 토론", CustomizeBoxType.TIME_BASED,
                                     null, null, 360, 120, null),
                             new CustomizeTimeBoxCreateRequest(Stance.NEUTRAL, "존중 토론", CustomizeBoxType.TIME_BASED,
@@ -230,8 +233,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                         fieldWithPath("table[].boxType").type(STRING).description("타임 박스 유형"),
                         fieldWithPath("table[].time").type(NUMBER).description("발언 시간(초)").optional(),
                         fieldWithPath("table[].bell").type(ARRAY).description("종소리 정보").optional(),
-                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)").optional(),
-                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수").optional(),
+                        fieldWithPath("table[].bell[].type").type(STRING).description("종소리 종류"),
+                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)"),
+                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수"),
                         fieldWithPath("table[].timePerTeam").type(NUMBER).description("팀당 발언 시간 (초)").optional(),
                         fieldWithPath("table[].timePerSpeaking").type(NUMBER).description("1회 발언 시간 (초)").optional(),
                         fieldWithPath("table[].speaker").type(STRING).description("발언자 이름").optional()
@@ -246,9 +250,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                             "찬성", "반대", true, true),
                     List.of(
                             new CustomizeTimeBoxResponse(Stance.PROS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellResponse(90, 1)), null, null, "콜리"),
+                                    120, List.of(new BellResponse(BellType.AFTER_START, 90, 1)), null, null, "콜리"),
                             new CustomizeTimeBoxResponse(Stance.CONS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellResponse(90, 1), new BellResponse(120, 2)), null, null, "비토"),
+                                    120, List.of(new BellResponse(BellType.AFTER_START, 90, 1), new BellResponse(BellType.AFTER_START, 120, 2)), null, null, "비토"),
                             new CustomizeTimeBoxResponse(Stance.NEUTRAL, "난상 토론", CustomizeBoxType.TIME_BASED,
                                     null, null, 360, 120, null),
                             new CustomizeTimeBoxResponse(Stance.NEUTRAL, "존중 토론", CustomizeBoxType.TIME_BASED,
@@ -323,8 +327,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                         fieldWithPath("table[].boxType").type(STRING).description("타임 박스 유형"),
                         fieldWithPath("table[].time").type(NUMBER).description("발언 시간(초)").optional(),
                         fieldWithPath("table[].bell").type(ARRAY).description("종소리 정보").optional(),
-                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)").optional(),
-                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수").optional(),
+                        fieldWithPath("table[].bell[].type").type(STRING).description("종소리 종류"),
+                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)"),
+                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수"),
                         fieldWithPath("table[].timePerTeam").type(NUMBER).description("팀당 발언 시간 (초)").optional(),
                         fieldWithPath("table[].timePerSpeaking").type(NUMBER).description("1회 발언 시간 (초)").optional(),
                         fieldWithPath("table[].speaker").type(STRING).description("발언자 이름").optional()
@@ -347,8 +352,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                         fieldWithPath("table[].boxType").type(STRING).description("타임 박스 유형"),
                         fieldWithPath("table[].time").type(NUMBER).description("발언 시간(초)").optional(),
                         fieldWithPath("table[].bell").type(ARRAY).description("종소리 정보").optional(),
-                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)").optional(),
-                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수").optional(),
+                        fieldWithPath("table[].bell[].type").type(STRING).description("종소리 종류"),
+                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)"),
+                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수"),
                         fieldWithPath("table[].timePerTeam").type(NUMBER).description("팀당 발언 시간 (초)").optional(),
                         fieldWithPath("table[].timePerSpeaking").type(NUMBER).description("1회 발언 시간 (초)").optional(),
                         fieldWithPath("table[].speaker").type(STRING).description("발언자 이름").optional()
@@ -362,9 +368,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                             "반대", true, true),
                     List.of(
                             new CustomizeTimeBoxCreateRequest(Stance.PROS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellRequest(90, 1)), null, null, "콜리"),
+                                    120, List.of(new BellRequest(BellType.AFTER_START, 90, 1)), null, null, "콜리"),
                             new CustomizeTimeBoxCreateRequest(Stance.CONS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellRequest(90, 1), new BellRequest(120, 2)), null, null, "비토"),
+                                    120, List.of(new BellRequest(BellType.AFTER_START, 90, 1), new BellRequest(BellType.AFTER_START, 120, 2)), null, null, "비토"),
                             new CustomizeTimeBoxCreateRequest(Stance.NEUTRAL, "난상 토론", CustomizeBoxType.TIME_BASED,
                                     null, null, 360, 120, null),
                             new CustomizeTimeBoxCreateRequest(Stance.NEUTRAL, "존중 토론", CustomizeBoxType.TIME_BASED,
@@ -377,9 +383,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                             "찬성", "반대", true, true),
                     List.of(
                             new CustomizeTimeBoxResponse(Stance.PROS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellResponse(90, 1)), null, null, "콜리"),
+                                    120, List.of(new BellResponse(BellType.AFTER_START, 90, 1)), null, null, "콜리"),
                             new CustomizeTimeBoxResponse(Stance.CONS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellResponse(90, 1), new BellResponse(120, 2)), null, null, "비토"),
+                                    120, List.of(new BellResponse(BellType.AFTER_START, 90, 1), new BellResponse(BellType.AFTER_START, 120, 2)), null, null, "비토"),
                             new CustomizeTimeBoxResponse(Stance.NEUTRAL, "난상 토론", CustomizeBoxType.TIME_BASED,
                                     null, null, 360, 120, null),
                             new CustomizeTimeBoxResponse(Stance.NEUTRAL, "존중 토론", CustomizeBoxType.TIME_BASED,
@@ -424,9 +430,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                             "반대", true, true),
                     List.of(
                             new CustomizeTimeBoxCreateRequest(Stance.PROS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellRequest(90, 1)), null, null, "콜리"),
+                                    120, List.of(new BellRequest(BellType.AFTER_START, 90, 1)), null, null, "콜리"),
                             new CustomizeTimeBoxCreateRequest(Stance.CONS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellRequest(90, 1), new BellRequest(120, 2)), null, null, "비토"),
+                                    120, List.of(new BellRequest(BellType.AFTER_START, 90, 1), new BellRequest(BellType.AFTER_START, 120, 2)), null, null, "비토"),
                             new CustomizeTimeBoxCreateRequest(Stance.NEUTRAL, "난상 토론", CustomizeBoxType.TIME_BASED,
                                     null, null, 360, 120, null),
                             new CustomizeTimeBoxCreateRequest(Stance.NEUTRAL, "존중 토론", CustomizeBoxType.TIME_BASED,
@@ -488,8 +494,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                         fieldWithPath("table[].boxType").type(STRING).description("타임 박스 유형"),
                         fieldWithPath("table[].time").type(NUMBER).description("발언 시간(초)").optional(),
                         fieldWithPath("table[].bell").type(ARRAY).description("종소리 정보").optional(),
-                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)").optional(),
-                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수").optional(),
+                        fieldWithPath("table[].bell[].type").type(STRING).description("종소리 종류"),
+                        fieldWithPath("table[].bell[].time").type(NUMBER).description("종소리 울릴 시간(초)"),
+                        fieldWithPath("table[].bell[].count").type(NUMBER).description("종소리 횟수"),
                         fieldWithPath("table[].timePerTeam").type(NUMBER).description("팀당 발언 시간 (초)").optional(),
                         fieldWithPath("table[].timePerSpeaking").type(NUMBER).description("1회 발언 시간 (초)").optional(),
                         fieldWithPath("table[].speaker").type(STRING).description("발언자 이름").optional()
@@ -505,9 +512,9 @@ public class CustomizeDocumentTest extends BaseDocumentTest {
                             "찬성", "반대", true, true),
                     List.of(
                             new CustomizeTimeBoxResponse(Stance.PROS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellResponse(90, 1)), null, null, "콜리"),
+                                    120, List.of(new BellResponse(BellType.AFTER_START, 90, 1)), null, null, "콜리"),
                             new CustomizeTimeBoxResponse(Stance.CONS, "입론", CustomizeBoxType.NORMAL,
-                                    120, List.of(new BellResponse(90, 1), new BellResponse(120, 2)), null, null, "비토"),
+                                    120, List.of(new BellResponse(BellType.AFTER_START, 90, 1), new BellResponse(BellType.AFTER_START, 120, 2)), null, null, "비토"),
                             new CustomizeTimeBoxResponse(Stance.NEUTRAL, "난상 토론", CustomizeBoxType.TIME_BASED,
                                     null, null, 360, 120, null),
                             new CustomizeTimeBoxResponse(Stance.NEUTRAL, "존중 토론", CustomizeBoxType.TIME_BASED,
