+++ src/test/java/com/debatetimer/domain/customize/BellTest.java
@@ -17,21 +17,21 @@ class BellTest {
 
         @Test
         void 벨_시간이_음수면_생성되지_않는다() {
-            assertThatThrownBy(() -> new Bell(-1, 1))
+            assertThatThrownBy(() -> new Bell(BellType.AFTER_START, -1, 1))
                     .isInstanceOf(DTClientErrorException.class)
                     .hasMessage(ClientErrorCode.INVALID_BELL_TIME.getMessage());
         }
 
         @Test
         void 벨_시간은_0이상이어야_한다() {
-            assertThatCode(() -> new Bell(0, 1))
+            assertThatCode(() -> new Bell(BellType.AFTER_START, 0, 1))
                     .doesNotThrowAnyException();
         }
 
         @ValueSource(ints = {0, Bell.MAX_BELL_COUNT + 1})
         @ParameterizedTest
         void 벨_횟수는_정해진_횟수_바깥일_경우_생성되지_않는다(int count) {
-            assertThatThrownBy(() -> new Bell(1, count))
+            assertThatThrownBy(() -> new Bell(BellType.AFTER_START, 1, count))
                     .isInstanceOf(DTClientErrorException.class)
                     .hasMessage(ClientErrorCode.INVALID_BELL_COUNT.getMessage());
         }
@@ -39,7 +39,7 @@ class BellTest {
         @ValueSource(ints = {1, Bell.MAX_BELL_COUNT})
         @ParameterizedTest
         void 벨_횟수는_정해진_횟수_이내여야_한다(int count) {
-            assertThatCode(() -> new Bell(1, count))
+            assertThatCode(() -> new Bell(BellType.AFTER_START, 1, count))
                     .doesNotThrowAnyException();
         }
     }
