+++ src/test/java/com/debatetimer/entity/customize/BellEntityTest.java
@@ -2,6 +2,7 @@ package com.debatetimer.entity.customize;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.debatetimer.domain.customize.BellType;
 import com.debatetimer.exception.custom.DTClientErrorException;
 import com.debatetimer.exception.errorcode.ClientErrorCode;
 import org.junit.jupiter.api.Nested;
@@ -16,7 +17,7 @@ class BellEntityTest {
 
         @Test
         void 벨_시간은_0이상이어야_한다() {
-            assertThatThrownBy(() -> new BellEntity(null, -1, 1))
+            assertThatThrownBy(() -> new BellEntity(null, BellType.AFTER_END, -1, 1))
                     .isInstanceOf(DTClientErrorException.class)
                     .hasMessage(ClientErrorCode.INVALID_BELL_TIME.getMessage());
         }
@@ -24,7 +25,7 @@ class BellEntityTest {
         @ValueSource(ints = {0, BellEntity.MAX_BELL_COUNT + 1})
         @ParameterizedTest
         void 벨_횟수는_정해진_횟수_이내여야_한다(int count) {
-            assertThatThrownBy(() -> new BellEntity(null, 1, count))
+            assertThatThrownBy(() -> new BellEntity(null, BellType.AFTER_END, 1, count))
                     .isInstanceOf(DTClientErrorException.class)
                     .hasMessage(ClientErrorCode.INVALID_BELL_COUNT.getMessage());
         }
