+++ src/test/java/com/debatetimer/fixture/BellGenerator.java
@@ -1,5 +1,6 @@
 package com.debatetimer.fixture;
 
+import com.debatetimer.domain.customize.BellType;
 import com.debatetimer.entity.customize.BellEntity;
 import com.debatetimer.entity.customize.CustomizeTimeBoxEntity;
 import com.debatetimer.repository.customize.BellRepository;
@@ -14,8 +15,8 @@ public class BellGenerator {
         this.bellRepository = bellRepository;
     }
 
-    public BellEntity generate(CustomizeTimeBoxEntity timeBox, int time, int count) {
-        BellEntity bell = new BellEntity(timeBox, time, count);
+    public BellEntity generate(CustomizeTimeBoxEntity timeBox, BellType type, int time, int count) {
+        BellEntity bell = new BellEntity(timeBox, type, time, count);
         return bellRepository.save(bell);
     }
 }
