+++ src/test/java/com/debatetimer/repository/customize/BellRepositoryTest.java
@@ -3,6 +3,7 @@ package com.debatetimer.repository.customize;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import com.debatetimer.domain.customize.BellType;
 import com.debatetimer.domain.customize.CustomizeBoxType;
 import com.debatetimer.domain.member.Member;
 import com.debatetimer.entity.customize.CustomizeTableEntity;
@@ -30,9 +31,9 @@ class BellRepositoryTest extends BaseRepositoryTest {
                     CustomizeBoxType.NORMAL, 1);
             CustomizeTimeBoxEntity otherTimeBox = customizeTimeBoxGenerator.generate(otherTable,
                     CustomizeBoxType.NORMAL, 1);
-            bellGenerator.generate(deleteBellTimeBox, 45, 1);
-            bellGenerator.generate(deleteBellTimeBox, 60, 1);
-            bellGenerator.generate(otherTimeBox, 45, 1);
+            bellGenerator.generate(deleteBellTimeBox, BellType.AFTER_START, 45, 1);
+            bellGenerator.generate(deleteBellTimeBox, BellType.AFTER_START, 60, 1);
+            bellGenerator.generate(otherTimeBox, BellType.AFTER_START, 45, 1);
 
             bellRepository.deleteAllByTable(deleteBellTable.getId());
